+++ include/bitcoin/bitcoin/coinninja/transaction/unspent_transaction_output.hpp
@@ -32,13 +32,13 @@ public:
     /**
      * Constructors
      */
-    unspent_transaction_output(std::string txid, uint8_t index, uint64_t amount, coinninja::wallet::derivation_path path, bool is_confirmed);
+    unspent_transaction_output(std::string txid, uint32_t index, uint64_t amount, coinninja::wallet::derivation_path path, bool is_confirmed);
 
     /**
      * Members
      */
     std::string txid;
-    uint8_t index;
+    uint32_t index;
     uint64_t amount;
     coinninja::wallet::derivation_path path;
     bool is_confirmed;
